chore: move transfer to measurement structure
