+++ front/src/pages/measurements/transfer/detail/B1505Transfer.tsx
