+++ front/src/pages/measurements/transfer/detail/index.tsx
