+++ front/src/pages/measurements/transfer/list/MeasurementPlot/MeasurementTypeRender.tsx
